% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/droppr_analyze.R
\name{droppr_analyze}
\alias{droppr_analyze}
\title{Analyze results from \code{droppr_main}}
\usage{
droppr_analyze(droppr.out)
}
\arguments{
\item{droppr.out}{an object of class "droppr"}
}
\value{
Returns a named list with method-specific components.
}
\description{
Implements several method-specific routines for clarifying
results from \code{droppr_main}
}

