% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/droppr_dbscan_data.R
\name{droppr_dbscan_data}
\alias{droppr_dbscan_data}
\title{Prepare dataset for \code{droppr} with DBSCAN}
\usage{
droppr_dbscan_data(d, eps, min.pts, ...)
}
\arguments{
\item{d}{a data frame}

\item{eps}{reachability distance}

\item{min.pts}{minimum number of points}

\item{...}{optional arguments passed to \code{fpc::dbscan()}}
}
\value{
Returns a data frame with a column of DBSCAN clusters
}
\description{
Adds a column of DBSCAN clusters to a data frame
}

