% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/droppr_dbscan_nclust.R
\name{droppr_dbscan_nclust}
\alias{droppr_dbscan_nclust}
\title{Identify optimal DBSCAN parameters}
\usage{
droppr_dbscan_nclust(d, eps, min.pts, ...)
}
\arguments{
\item{d}{a data frame}

\item{eps}{a vector of reachability distances}

\item{min.pts}{a vector of minimum points values}

\item{...}{optional arguments passed to \code{fpc::dbscan()}}
}
\value{
Returns a named list with the following components:

\item{\code{params.df}}{a data frame containing combinations of parameter values and the resulting number of DBSCAN clusters}
\item{\code{params.plot}}{a ggplot object depicting the elbow curve for parameter combinations}
}
\description{
Iterates DBSCAN over various combinations of parameter values
and returns the associated number of clusters
}

